Made button class, implemented tags
Decided to go with a tags system to allow other objects (enemies) to find objects with a specific tag (player). Not extremely efficient as you have to loop over the whole GameObjects vector and check each tag but I will counter this by putting the player at the start of the vector. Plus the game shouldn't have THAT many GameObjects so it'll be fine (probably). Also changed UI objects to have their transform position be the centre instead of the bottom left making positioning them a lot more simple. Also put the openGL wrapper classes in a separate folder. This time I'm actually close to making starting to make a game (I think).