DeltaTime calculated from inside Game
To be able to set deltaTime to 0 when changing a scene as we want the scene to load and then deltaTime to start. Also changed SetUniformValues to  UseCamera so you can pass in a camera which will have the uniform values set according to that. Will help with multiple cameras in the future. Also easier to understand name.