Implemented UI anchors
Also decided on a system to keep track of GameObjects. Use an unordered map to keep track of any named GameObject and then add the GameObjects that I want to be able to find again to that unordered map as well as the general GameObjects vector. This allows for simpler code in terms of updating/drawing as I only need to iterate through the mGameObjects vector. Very nice.