Set up camera to follow player, started Enemy
Also changed the camera so that it looks "backwards" which is actually the way it's facing. Added GetUp() and GetRight() to transform for future use. Rotation of objects might become a problem due to not using quaternion, the problem comes from just rotation around one axis at a time. I ran into that problem with the camera but I could bypass it for now by just rotating around the y axis first and then the x axis, but if I needed to rotate around all 3 axis then I may have to move to a quaternion system.