Got rid of lazy loading
Lazy loading useful if you don't know if you're going to initialise OpenGL before or after you OpenGL classes, but that's not our case. For us it just makes the game stutter when a texture/model/shader is being used for the first time. Now classes send their data to the GPU when they're initialised. Since we're using a model library type architecture, all textures/models/shaders will be initialised at the start when the Game class is meaning a longer initial load but I think that's worth it for a smoother running game. Also got rid of unnecessary initialisation lists because I think they look bad. Also wasn't able to get rid of lazy loading in the mesh class because the way it works is by adding faces to the mesh after the mesh has been initialised. Could have sent data to the GPU every time a face is added but many faces would cause a long initial load time so we'll just take the hit at runtime instead.