Implemented text rendering
Long day. Thanks to https://learnopengl.com/In-Practice/Text-Rendering and after rearranging some code I was able to render text. I wrapped it into it's own text class and have included it on the buttons in the menu and as a game speed display. I've seen that the given code might not be the most efficient so I may come back to it (can't wait!)