Fixed text rendering
Text would either render in front of other ui but behind 3D objects with depth testing off, or behind ui and in front of 3D objects but with the transparent space also blocking the objects behind. Fixed by setting gl_FragDepth manually in both font and ui fragment shaders and leaving depth testing on so we can dictate what gets drawn in front in terms of ui. Also added a test if the alpha value of the fragment is very low then we discard it fixing the transparent space issue. Also went and tidied up the shaders, tabbing lines correctly and making uniform names all correlate with each other.