Went back to lazy loading
Went on holiday so no development for a while (:disappointed:) and had the epiphany that OpenGL uploads data to the GPU so we want to save vram by not uploading textures/models if we don't need them. This then means that we are leaving the data on the GPU after the texture/model has been used which isn't optimal (vram will only go up, never down). I had an idea with shaders using shared pointers to track the number of instances so that we can determine if there aren't any objects holding a reference to the shader (and therefore not using the shader) then we can not bother sending uniform information to that shader. The same might be applicable with models/textures but it would be more important here. On my machine the game gets 3500fps (if unlimited) and uses around 200mb of ram, but I think I'll try to add the optimisations in case I try to use this frame work to make a more demanding game (and the task sounds fun).
Also removed large skybox texture from assets (so I can send the game to my friend), made the menu have a skybox and the gameplay scene just have a solid background, and added an fps counter in the gameplay scene.