Added grass
Also tried to fix the fps limit which I swear was working last time, not only working but working well. The actual frame rate was very inconsistent and was higher than what we wanted. By moving the delay after the update and draw calls it fixes it a bit but it's still fairly inconsistent.