Implemented shared pointers for Shaders
Turned Shaders, stored in ShaderLibrary, into shared pointers along with holding them as shared pointers in the 3D or UI object. This was so that I could find how many objects were in the scene and using the shaders so I could determine if I needed to set uniforms for them or not. Contemplated just leaving them as shared pointers come with some overhead and the Shaders weren't even pointers so they were stored on the stack, But ultimately decided that a much bigger project with much more complicated and expensive shaders would benefit from it, plus I don't have much hands on experience with smart pointers. I think I will do the same with Textures and Models and attempt to unload them from the GPU if they are not being used, because right now once they're loaded they're never unloaded, which again is fine for my small game but wouldn't be ideal in a much bigger game.