Implemented shared pointers for models
And meshes and textures stored in ModelLibrary. I also added an unload function to each OpenGL object that unloads it from the GPU. I added an update function for the ModelLibrary that checks how many active instances of the model/mesh/texture there is to determine if it can be unloaded. I am struggling to test that this really works though. I may need a GPU profiling tool to look at the memory. But it SHOULD work.